Add files via upload
OELMS is an online system developed using html, css,  js and google fonts. This is used by students to attend virtual classes, review and rate lectures, and get in touch with their tutors. On other hand, tutors use the system to upload learning materials, hold virtual classes and give out assignments. The content uploaded to the site is reviewed by the admin. The system is interactive and responsive to users